fixes line item duplication and quantity
We were using the line item id previously, but cannot because each
return holds one item, and in the case of multiple quantities of an
item returned, we had the issue of duplicated line items ids.
The fix has been to use an incremental index for line numbers,
and making sure that quantity was always 1.