Add InventorySummary document
The QL API returns InventorySummary documents with Blackboard if set up by the client or an InventorySummaryRequest message was sent upfront.

Unfortunately and contrary to their documentation the DocumentType returned inside the EventMessage is InventorySummaryReady. So, that is why the document is called InventorySummaryReady instead of InventorySummary (although the XMl still is a InventorySummary).